Split solvers.py in several files
